requirements.txt
asgiref==3.4.1
astroid==2.6.2
autopep8==1.5.7
coverage==5.5
Django==3.2.5
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
phonenumbers==8.12.27
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
Pygments==2.9.0
pylint==2.9.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
